refactored the ChangeRequestManager to not have the unused interface
OCD-3122
