+++ chpl/chpl-service/src/test/java/gov/healthit/chpl/manager/ChangeRequestManagerImplTest.java
@@ -22,7 +22,7 @@ import gov.healthit.chpl.changerequest.domain.service.ChangeRequestDetailsFactor
 import gov.healthit.chpl.changerequest.domain.service.ChangeRequestDetailsService;
 import gov.healthit.chpl.changerequest.domain.service.ChangeRequestStatusService;
 import gov.healthit.chpl.changerequest.manager.ChangeRequestManager;
-import gov.healthit.chpl.changerequest.manager.ChangeRequestManagerImpl;
+import gov.healthit.chpl.changerequest.manager.ChangeRequestManager;
 import gov.healthit.chpl.changerequest.validation.ChangeRequestValidationFactory;
 import gov.healthit.chpl.changerequest.validation.WebsiteValidation;
 import gov.healthit.chpl.domain.CertificationBody;
@@ -41,7 +41,7 @@ public class ChangeRequestManagerImplTest {
         Mockito.when(changeRequestDAO.get(ArgumentMatchers.anyLong()))
                 .thenReturn(getBasicChangeRequest());
 
-        ChangeRequestManager changeRequestManager = new ChangeRequestManagerImpl(changeRequestDAO,
+        ChangeRequestManager changeRequestManager = new ChangeRequestManager(changeRequestDAO,
                 null, null, null, null, null, null, null, null, null, null, null);
 
         // Run
@@ -60,7 +60,7 @@ public class ChangeRequestManagerImplTest {
         Mockito.when(changeRequestDAO.get(ArgumentMatchers.anyLong()))
                 .thenThrow(EntityRetrievalException.class);
 
-        ChangeRequestManager changeRequestManager = new ChangeRequestManagerImpl(changeRequestDAO,
+        ChangeRequestManager changeRequestManager = new ChangeRequestManager(changeRequestDAO,
                 null, null, null, null, null, null, null, null, null, null, null);
 
         // Run
@@ -77,7 +77,7 @@ public class ChangeRequestManagerImplTest {
         Mockito.when(changeRequestDAO.getAll())
                 .thenReturn(Arrays.asList(getBasicChangeRequest(), getBasicChangeRequest(), getBasicChangeRequest()));
 
-        ChangeRequestManager changeRequestManager = new ChangeRequestManagerImpl(changeRequestDAO,
+        ChangeRequestManager changeRequestManager = new ChangeRequestManager(changeRequestDAO,
                 null, null, null, null, null, null, null, null, null, null, null);
 
         // Run
@@ -110,7 +110,7 @@ public class ChangeRequestManagerImplTest {
         Mockito.when(crStatusService.updateChangeRequestStatus(ArgumentMatchers.any(ChangeRequest.class)))
                 .thenAnswer(i -> i.getArgument(0));
 
-        ChangeRequestManager changeRequestManager = new ChangeRequestManagerImpl(changeRequestDAO,
+        ChangeRequestManager changeRequestManager = new ChangeRequestManager(changeRequestDAO,
                 null,
                 null,
                 null,
@@ -144,7 +144,7 @@ public class ChangeRequestManagerImplTest {
         Mockito.when(websiteValidation.getMessages()).thenReturn(Arrays.asList("Error Message"));
         Mockito.when(crValidationFactory.getRule(ArgumentMatchers.anyString())).thenReturn(websiteValidation);
 
-        ChangeRequestManager changeRequestManager = new ChangeRequestManagerImpl(changeRequestDAO,
+        ChangeRequestManager changeRequestManager = new ChangeRequestManager(changeRequestDAO,
                 null,
                 null,
                 null,
@@ -187,7 +187,7 @@ public class ChangeRequestManagerImplTest {
         Mockito.when(crStatusService.updateChangeRequestStatus(ArgumentMatchers.any(ChangeRequest.class)))
                 .thenAnswer(i -> i.getArgument(0));
 
-        ChangeRequestManager changeRequestManager = new ChangeRequestManagerImpl(changeRequestDAO,
+        ChangeRequestManager changeRequestManager = new ChangeRequestManager(changeRequestDAO,
                 null,
                 null,
                 null,
@@ -231,7 +231,7 @@ public class ChangeRequestManagerImplTest {
         Mockito.when(crStatusService.updateChangeRequestStatus(ArgumentMatchers.any(ChangeRequest.class)))
                 .thenAnswer(i -> i.getArgument(0));
 
-        ChangeRequestManager changeRequestManager = new ChangeRequestManagerImpl(changeRequestDAO,
+        ChangeRequestManager changeRequestManager = new ChangeRequestManager(changeRequestDAO,
                 null,
                 null,
                 null,
